requirements.txt
alembic==1.13.0
astroid==3.0.2
attrs==23.1.0
bandit==1.7.5
bcrypt==4.1.1
black==23.11.0
blinker==1.7.0
click==8.1.7
dill==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Toastr==0.5.8
Flask-WTF==1.2.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
mypy==1.7.1
mypy-extensions==1.0.0
mysqlclient==2.2.0
packaging==23.2
pathspec==0.11.2
pbr==6.0.0
platformdirs==4.0.0
pluggy==1.3.0
pydot==2.0.0
Pygments==2.17.2
pylint==3.0.3
pyparsing==3.1.1
pytest==7.4.3
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.31.1
rich==13.7.0
rpds-py==0.13.2
ruff==0.1.6
setuptools==68.2.2
smmap==5.0.1
sql==2022.4.0
SQLAlchemy==2.0.23
sqlalchemy-schemadisplay==1.3
stevedore==5.1.0
tomlkit==0.12.3
typing_extensions==4.8.0
validators==0.22.0
Werkzeug==3.0.1
wheel==0.41.3
WTForms==3.1.1
